requirements.txt
alembic==1.4.2
astroid==2.3.3
click==7.1.1
dlib==19.19.0
dominate==2.5.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.2
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.15
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
